+++ .changeset/smooth-waves-complain.md
@@ -1,0 +1,7 @@
+---
+"@gouvfr-lasuite/proconnect.core": minor
+---
+
+👮 Accueillons l'équipe de sécurité de ProConnect
+
+Dans le cadres la migration du script d'import de comptes coop, une partie des fonctions de validation sont déplacées dans le package `@gouvfr-lasuite/proconnect.core/security` pour permettre leur réutilisation dans Hyyypertool.
