+++ packages/core/package.json
@@ -20,15 +20,15 @@
   "exports": {
     "./*": {
       "require": {
-        "types": "./dist/*",
-        "default": "./dist/*"
+        "types": "./dist/*/index.d.ts",
+        "default": "./dist/*/index.js"
       },
       "import": {
-        "types": "./dist/*",
-        "default": "./dist/*"
+        "types": "./dist/*/index.d.ts",
+        "default": "./dist/*/index.js"
       },
-      "types": "./dist/*",
-      "default": "./dist/*"
+      "types": "./dist/*/index.d.ts",
+      "default": "./dist/*/index.js"
     }
   },
   "scripts": {
@@ -46,14 +46,19 @@
   },
   "dependencies": {
     "@zootools/email-spell-checker": "^1.12.0",
-    "is-disposable-email-domain": "^1.0.7"
+    "is-disposable-email-domain": "^1.0.7",
+    "lodash-es": "^4.17.21",
+    "nanoid": "^5.0.9",
+    "tld-extract": "^2.1.0"
   },
   "devDependencies": {
     "@tsconfig/node22": "^22.0.0",
-    "@types/mocha": "^10.0.7",
-    "chai": "^5.1.1",
-    "mocha": "^10.7.3",
-    "tsx": "^4.17.0"
+    "@types/lodash-es": "^4.17.12",
+    "@types/mocha": "^10.0.10",
+    "@types/node": "^22.10.2",
+    "chai": "^5.1.2",
+    "mocha": "^11.0.1",
+    "tsx": "^4.19.2"
   },
   "publishConfig": {
     "access": "public",
