+++ packages/core/src/security/generate-pin-token.ts
@@ -1,0 +1,13 @@
+//
+
+import { customAlphabet } from "nanoid";
+
+//
+
+const nanoidPin = customAlphabet("0123456789", 10);
+
+//
+
+export function generatePinToken() {
+  return nanoidPin();
+}
