+++ packages/core/src/services/email/index.ts
@@ -1,0 +1,3 @@
+//
+
+export * from "./isAFreeDomain.js";
