+++ packages/core/src/services/suggestion/index.ts
@@ -1,0 +1,3 @@
+//
+
+export * from "./did-you-mean.js";
