+++ packages/core/tsconfig.json
@@ -5,7 +5,7 @@
     "declarationMap": true,
     "outDir": "./dist",
     "rootDir": "src",
-    "types": ["./types"],
+    "types": ["node", "./types"],
     "module": "NodeNext",
     "moduleResolution": "nodenext",
     "verbatimModuleSyntax": true,
