+++ packages/core/types/tld-extract.d.ts
@@ -1,0 +1,5 @@
+//
+
+declare module "tld-extract" {
+  function parse_host(domain: string, { allowDotlessTLD: boolean }): boolean;
+}
