+++ scripts/import-domains.ts
@@ -1,4 +1,8 @@
 // src https://stackoverflow.com/questions/40994095/pipe-streams-to-edit-csv-file-in-node-js
+import {
+  isDomainValid,
+  isSiretValid,
+} from "@gouvfr-lasuite/proconnect.core/security";
 import { AxiosError } from "axios";
 import { parse, stringify, transform } from "csv";
 import fs from "fs";
@@ -23,7 +27,6 @@ import {
   startDurationMesure,
   throttleApiCall,
 } from "../src/services/script-helpers";
-import { isDomainValid, isSiretValid } from "../src/services/security";
 
 const { INPUT_FILE, OUTPUT_FILE } = z
   .object({
