--- src/types/is-disposable-email-domain.d.ts
@@ -1,1 +1,0 @@
-declare module "is-disposable-email-domain";
